Add HomePage component
