+++ client/src/pages/HomePage/index.js
@@ -1,0 +1,3 @@
+import HomePage from './HomePage'
+
+export default HomePage;
